fix: enforce non-zero priority fee on op stack networks
